Change the way :free definition of numbers is done when the literal is in the same base as the destination number for consistency.
Formerly, when the input base was the same as the input base when constructing
a number from a literal and the :short mode was used:
  Num[base].new(literal, :short, base: base)
the result was equivalent to using the :free mode and preserved all the digits
of the input. Now the result is consistent with the different bases case, and
the result has just enough digits to guarantee conversion back to
with :fixed precision.